+++ docs/user-guide/run-benchmark.md
@@ -114,19 +114,18 @@ genai-bench benchmark --api-backend oci-cohere \
 
 ### Start a benchmark against OCI Cohere
 
-Below is a sample command to trigger a benchmark against cohere chat API.
+Below is a sample command to trigger a benchmark against Cohere chat API.
 
-> **Command-A models**: Add `--oci-cohere-api-version v2` when targeting Cohere Command-A reasoning or vision deployments on OCI. The V2 API unlocks reasoning traces and multimodal prompts.
+> **Command-A models**: Use `--api-backend oci-cohere-v2` when targeting Cohere Command-A reasoning or vision deployments on OCI. The V2 API unlocks reasoning traces and multimodal prompts.
 
 ```shell
-genai-bench benchmark --api-backend oci-cohere \
+genai-bench benchmark --api-backend oci-cohere-v2 \
             --config-file /home/ubuntu/.oci/config \
             --api-base "https://inference.generativeai.us-chicago-1.oci.oraclecloud.com" \
             --api-model-name "c4ai-command-r-08-2024" \
             --model-tokenizer "/home/ubuntu/c4ai-command-r-08-2024" \
             --server-engine "SGLang" \
             --task text-to-text \
-            --oci-cohere-api-version v2 \
             --num-concurrency 1 \
             --server-gpu-type A100-80G \
             --server-version "command_r_082024_v1_7" \
@@ -208,13 +207,12 @@ Model-specific parameters like `quality`, `response_format`, `output_format`, or
 ### Benchmark a Command-A vision model on OCI Cohere
 
 ```shell
-genai-bench benchmark --api-backend oci-cohere \
+genai-bench benchmark --api-backend oci-cohere-v2 \
             --config-file /home/ubuntu/.oci/config \
             --api-base "https://inference.generativeai.us-chicago-1.oci.oraclecloud.com" \
             --api-model-name "cohere.command-a-vision" \
             --model-tokenizer "/home/ubuntu/cohere-command-a-tokenizer" \
             --task image-text-to-text \
-            --oci-cohere-api-version v2 \
             --dataset-path /home/ubuntu/datasets/vision.jsonl \
             --num-concurrency 1 \
             --server-gpu-type A100-80G \
