+++ tests/cli/test_cli_benchmark.py
@@ -26,7 +26,6 @@ import pytest
 from click.testing import CliRunner
 
 from genai_bench.cli.cli import benchmark
-from genai_bench.user.oci_cohere_user import OCICohereUser
 
 
 @pytest.fixture
@@ -396,7 +395,9 @@ def test_benchmark_cohere_v2_rejects_embeddings_task(cli_runner, default_options
         config_file.write(config.encode("utf-8"))
         config_file.flush()
 
-        options = [opt if opt != "openai" else "oci-cohere" for opt in default_options]
+        options = [
+            opt if opt != "openai" else "oci-cohere-v2" for opt in default_options
+        ]
         options = [
             opt if opt != "text-to-text" else "text-to-embeddings" for opt in options
         ]
@@ -407,14 +408,12 @@ def test_benchmark_cohere_v2_rejects_embeddings_task(cli_runner, default_options
                 *options,
                 "--config-file",
                 config_file.name,
-                "--oci-cohere-api-version",
-                "v2",
             ],
         )
     assert result.exit_code != 0, (
         "Expected failure for v2 + text-to-embeddings but command succeeded"
     )
-    assert "supports chat benchmarks only" in result.output
+    assert "Task 'text-to-embeddings' is not supported" in result.output
 
 
 @pytest.mark.usefixtures(
@@ -429,42 +428,34 @@ def test_benchmark_cohere_v2_rejects_embeddings_task(cli_runner, default_options
     "mock_experiment_path",
 )
 def test_benchmark_command_with_cohere_v2(cli_runner, default_options):
-    original_version = OCICohereUser.api_version
-    OCICohereUser.api_version = "v1"
-    try:
-        with (
-            tempfile.NamedTemporaryFile() as config_file,
-            tempfile.NamedTemporaryFile() as key_file,
-        ):
-            key_file.write(b"key")
-            key_file.flush()
-            config = f"""[DEFAULT]
-    user=ocid1.user.oc1..example
-    fingerprint=8c:4c:01:71:6e:4a:e8:11:ab:c9:c2:63:fb:36:f5:ec
-    tenancy=ocid1.tenancy.oc1..example
-    region=us-ashburn-1
-    key_file={key_file.name}"""
-            config_file.write(config.encode("utf-8"))
-            config_file.flush()
+    with (
+        tempfile.NamedTemporaryFile() as config_file,
+        tempfile.NamedTemporaryFile() as key_file,
+    ):
+        key_file.write(b"key")
+        key_file.flush()
+        config = f"""[DEFAULT]
+user=ocid1.user.oc1..example
+fingerprint=8c:4c:01:71:6e:4a:e8:11:ab:c9:c2:63:fb:36:f5:ec
+tenancy=ocid1.tenancy.oc1..example
+region=us-ashburn-1
+key_file={key_file.name}"""
+        config_file.write(config.encode("utf-8"))
+        config_file.flush()
 
-            options = [
-                opt if opt != "openai" else "oci-cohere" for opt in default_options
-            ]
+        options = [
+            opt if opt != "openai" else "oci-cohere-v2" for opt in default_options
+        ]
 
-            result = cli_runner.invoke(
-                benchmark,
-                [
-                    *options,
-                    "--config-file",
-                    config_file.name,
-                    "--oci-cohere-api-version",
-                    "v2",
-                ],
-            )
-        assert result.exit_code == 0, f"Command failed with output: {result.output}"
-        assert OCICohereUser.api_version == "v2"
-    finally:
-        OCICohereUser.api_version = original_version
+        result = cli_runner.invoke(
+            benchmark,
+            [
+                *options,
+                "--config-file",
+                config_file.name,
+            ],
+        )
+    assert result.exit_code == 0, f"Command failed with output: {result.output}"
 
 
 @pytest.mark.usefixtures(
