+++ tests/cli/test_validation_multicloud.py
@@ -15,6 +15,7 @@ from genai_bench.user.aws_bedrock_user import AWSBedrockUser
 from genai_bench.user.azure_openai_user import AzureOpenAIUser
 from genai_bench.user.gcp_vertex_user import GCPVertexUser
 from genai_bench.user.oci_cohere_user import OCICohereUser
+from genai_bench.user.oci_cohere_v2_user import OCICohereV2User
 from genai_bench.user.openai_user import OpenAIUser
 
 
@@ -64,6 +65,13 @@ class TestMultiCloudValidation:
 
         assert result == "oci-cohere"
         assert self.ctx.obj["user_class"] == OCICohereUser
+
+    def test_validate_oci_cohere_v2_backend(self):
+        """Test OCI Cohere V2 backend validation."""
+        result = validate_api_backend(self.ctx, self.param, "oci-cohere-v2")
+
+        assert result == "oci-cohere-v2"
+        assert self.ctx.obj["user_class"] == OCICohereV2User
 
     def test_validate_vllm_backend(self):
         """Test vLLM backend validation (uses OpenAI user)."""
@@ -328,6 +336,7 @@ class TestMultiCloudValidation:
         expected_backends = [
             "openai",
             "oci-cohere",
+            "oci-cohere-v2",
             "cohere",
             "aws-bedrock",
             "azure-openai",
