+++ tests/user/test_oci_cohere_user.py
@@ -1,13 +1,9 @@
-import json
 from unittest.mock import MagicMock, patch
 
 import pytest
 from oci.generative_ai_inference.models import (
     ChatDetails,
     CohereChatRequest,
-    CohereChatRequestV2,
-    CohereImageContentV2,
-    CohereTextContentV2,
     DedicatedServingMode,
     EmbedTextDetails,
     OnDemandServingMode,
@@ -17,7 +13,6 @@ from oci.generative_ai_inference.models import (
 from genai_bench.protocol import (
     UserChatRequest,
     UserEmbeddingRequest,
-    UserImageChatRequest,
     UserImageEmbeddingRequest,
     UserReRankRequest,
 )
@@ -27,7 +22,6 @@ from genai_bench.user.oci_cohere_user import OCICohereUser
 @pytest.fixture
 def test_cohere_user():
     OCICohereUser.host = "http://example.com"
-    OCICohereUser.api_version = "v1"
 
     # Set up mock auth provider
     mock_auth = MagicMock()
@@ -40,259 +34,14 @@ def test_cohere_user():
     OCICohereUser.auth_provider = mock_auth
 
     user = OCICohereUser(environment=MagicMock())
-    try:
-        yield user
-    finally:
-        OCICohereUser.api_version = "v1"
+    yield user
 
 
-def test_supported_tasks_include_vision_chat():
-    assert OCICohereUser.supported_tasks["image-text-to-text"] == "chat"
+def test_supported_tasks_exclude_vision_chat():
+    assert "image-text-to-text" not in OCICohereUser.supported_tasks
 
 
-@patch("genai_bench.user.oci_cohere_user.GenerativeAiInferenceClient")
-def test_chat_v2_text(mock_client_class, test_cohere_user):
-    mock_client_instance = mock_client_class.return_value
-    mock_client_instance.chat.return_value.status = 200
-    mock_client_instance.chat.return_value.data.events.return_value = iter(
-        [
-            MagicMock(
-                data=json.dumps(
-                    {
-                        "message": {
-                            "role": "ASSISTANT",
-                            "content": [{"type": "TEXT", "text": "Hello"}],
-                        }
-                    }
-                ).encode("utf-8")
-            ),
-            MagicMock(
-                data=json.dumps(
-                    {
-                        "message": {
-                            "role": "ASSISTANT",
-                            "content": [
-                                {"type": "THINKING", "thinking": " Reason"},
-                            ],
-                        }
-                    }
-                ).encode("utf-8")
-            ),
-            MagicMock(
-                data=json.dumps(
-                    {
-                        "finishReason": "COMPLETE",
-                        "usage": {
-                            "promptTokens": 5,
-                            "completionTokens": 3,
-                            "completionTokensDetails": {
-                                "reasoningTokens": 2,
-                            },
-                        },
-                        "message": {
-                            "role": "ASSISTANT",
-                            "content": [{"type": "TEXT", "text": "!"}],
-                        },
-                    }
-                ).encode("utf-8")
-            ),
-        ]
-    )
-
-    test_cohere_user.api_version = "v2"
-    test_cohere_user.on_start()
-    history = [{"role": "assistant", "content": "Previous turn"}]
-    documents = [{"text": "Doc"}]
-    test_cohere_user.sample = lambda: UserChatRequest(
-        model="cohere-command-a",
-        prompt="Describe the data",
-        num_prefill_tokens=1,
-        additional_request_params={
-            "compartmentId": "ocid1.compartment.oc1..example",
-            "servingType": "ON_DEMAND",
-            "system_message": "Be concise",
-            "chat_history": history,
-            "documents": documents,
-            "top_p": 0.9,
-            "temperature": 0.3,
-        },
-        max_tokens=64,
-    )
-
-    metrics_collector_mock = MagicMock()
-    test_cohere_user.collect_metrics = metrics_collector_mock
-
-    test_cohere_user.chat()
-
-    chat_detail = mock_client_instance.chat.call_args[0][0]
-    chat_request = chat_detail.chat_request
-    assert isinstance(chat_request, CohereChatRequestV2)
-    assert chat_request.top_p == 0.9
-    assert chat_request.temperature == 0.3
-    assert chat_request.documents == documents
-    assert chat_request.stream_options.is_include_usage
-    assert chat_request.messages[0].role == "SYSTEM"
-    assert isinstance(chat_request.messages[0].content[0], CohereTextContentV2)
-
-    metrics_collector_mock.assert_called_once()
-    user_response = metrics_collector_mock.call_args[0][0]
-    assert user_response.tokens_received == 3
-    assert user_response.reasoning_tokens == 2
-    assert user_response.num_prefill_tokens == 5
-    assert user_response.status_code == 200
-
-
-@patch("genai_bench.user.oci_cohere_user.GenerativeAiInferenceClient")
-def test_chat_v2_text_snake_case_usage(mock_client_class, test_cohere_user):
-    mock_client_instance = mock_client_class.return_value
-    mock_client_instance.chat.return_value.status = 200
-    mock_client_instance.chat.return_value.data.events.return_value = iter(
-        [
-            MagicMock(
-                data=json.dumps(
-                    {
-                        "message": {
-                            "role": "ASSISTANT",
-                            "content": [{"type": "TEXT", "text": "Hi"}],
-                        }
-                    }
-                ).encode("utf-8")
-            ),
-            MagicMock(
-                data=json.dumps(
-                    {
-                        "finishReason": "COMPLETE",
-                        "usage": {
-                            "tokens": {
-                                "input_tokens": 4,
-                                "output_tokens": 2,
-                                "reasoning_tokens": 1,
-                            }
-                        },
-                        "message": {
-                            "role": "ASSISTANT",
-                            "content": [{"type": "TEXT", "text": " there"}],
-                        },
-                    }
-                ).encode("utf-8")
-            ),
-        ]
-    )
-
-    test_cohere_user.api_version = "v2"
-    test_cohere_user.on_start()
-    test_cohere_user.sample = lambda: UserChatRequest(
-        model="cohere-command-a",
-        prompt="Hi?",
-        num_prefill_tokens=None,
-        additional_request_params={
-            "compartmentId": "ocid1.compartment.oc1..example",
-            "servingType": "ON_DEMAND",
-        },
-        max_tokens=32,
-    )
-
-    metrics_collector_mock = MagicMock()
-    test_cohere_user.collect_metrics = metrics_collector_mock
-
-    test_cohere_user.chat()
-
-    metrics_collector_mock.assert_called_once()
-    user_response = metrics_collector_mock.call_args[0][0]
-    assert user_response.tokens_received == 2
-    assert user_response.reasoning_tokens == 1
-    assert user_response.num_prefill_tokens == 4
-
-
-@patch("genai_bench.user.oci_cohere_user.GenerativeAiInferenceClient")
-def test_chat_v2_vision(mock_client_class, test_cohere_user):
-    mock_client_instance = mock_client_class.return_value
-    mock_client_instance.chat.return_value.status = 200
-    mock_client_instance.chat.return_value.data.events.return_value = iter(
-        [
-            MagicMock(
-                data=json.dumps(
-                    {
-                        "message": {
-                            "role": "ASSISTANT",
-                            "content": [{"type": "TEXT", "text": "Analyzing"}],
-                        }
-                    }
-                ).encode("utf-8")
-            ),
-            MagicMock(
-                data=json.dumps(
-                    {
-                        "finishReason": "COMPLETE",
-                        "message": {
-                            "role": "ASSISTANT",
-                            "content": [{"type": "TEXT", "text": " image"}],
-                        },
-                    }
-                ).encode("utf-8")
-            ),
-        ]
-    )
-
-    test_cohere_user.api_version = "v2"
-    test_cohere_user.on_start()
-    image_data = "data:image/png;base64,AAA"
-    test_cohere_user.sample = lambda: UserImageChatRequest(
-        model="cohere-command-a-vision",
-        prompt="What do you see?",
-        image_content=[image_data],
-        num_images=1,
-        num_prefill_tokens=None,
-        additional_request_params={
-            "compartmentId": "ocid1.compartment.oc1..example",
-            "servingType": "ON_DEMAND",
-        },
-        max_tokens=32,
-    )
-
-    metrics_collector_mock = MagicMock()
-    test_cohere_user.collect_metrics = metrics_collector_mock
-
-    test_cohere_user.chat()
-
-    chat_detail = mock_client_instance.chat.call_args[0][0]
-    chat_request = chat_detail.chat_request
-    user_message = chat_request.messages[-1]
-    image_contents = [
-        content
-        for content in user_message.content
-        if isinstance(content, CohereImageContentV2)
-    ]
-    assert len(image_contents) == 1
-    assert image_contents[0].image_url.url == image_data
-    metrics_collector_mock.assert_called_once()
-
-
-@patch("genai_bench.user.oci_cohere_user.GenerativeAiInferenceClient")
-def test_chat_v2_vision_multiple_images_error(mock_client_class, test_cohere_user):
-    mock_client_class.return_value.chat.return_value.status = 200
-
-    test_cohere_user.api_version = "v2"
-    test_cohere_user.on_start()
-    test_cohere_user.sample = lambda: UserImageChatRequest(
-        model="cohere-command-a-vision",
-        prompt="Describe both images",
-        image_content=["img1", "img2"],
-        num_images=2,
-        num_prefill_tokens=None,
-        additional_request_params={
-            "compartmentId": "ocid1.compartment.oc1..example",
-            "servingType": "ON_DEMAND",
-        },
-        max_tokens=32,
-    )
-
-    with pytest.raises(ValueError, match="only a single image"):
-        test_cohere_user.chat()
-
-
-def test_embeddings_use_legacy_parse_with_v2_enabled(test_cohere_user):
-    test_cohere_user.api_version = "v2"
+def test_embeddings_use_legacy_parse(test_cohere_user):
     mock_send = MagicMock()
     test_cohere_user.send_request = mock_send
     test_cohere_user.sample = lambda: UserEmbeddingRequest(
@@ -313,8 +62,7 @@ def test_embeddings_use_legacy_parse_with_v2_enabled(test_cohere_user):
     )
 
 
-def test_rerank_use_legacy_parse_with_v2_enabled(test_cohere_user):
-    test_cohere_user.api_version = "v2"
+def test_rerank_use_legacy_parse(test_cohere_user):
     mock_send = MagicMock()
     test_cohere_user.send_request = mock_send
     test_cohere_user.sample = lambda: UserReRankRequest(
